LayoutTests/platform/glib/http/tests/inspector/network/intercept-request-main-resource-expected.txt

ALERT: ORIGINAL HTML CONTENT
FAIL: Timed out waiting for notifyDone to be called

Bad Request

Your browser sent a request that this server could not understand.

Apache/2.4.37 (Unix) Server at 127.0.0.1 Port 80
